\name{CovControl-class}
\docType{class}
\alias{CovControl-class}

\title{Class "CovControl" is a VIRTUAL base control class }
\description{  The class "CovControl" is a VIRTUAL base control class 
    for the derived classes representing the control parameters 
    for the different robust methods }
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "CovControl" in the signature.
}
%\references{}
\author{Valentin Todorov \email{valentin.todorov@chello.at}}
%\note{}

%\seealso{}
%\examples{}
\keyword{classes}
\keyword{robust}
\keyword{multivariate}
